Aggressively print STEP_CURSOR annotations.
When steps have substeps -- e.g. when a step runs a script that itself emits
annotations circumventing the StructuredAnnotationStream machinery -- it is
likely that the, once all of those substeps have completed, the step cursor will
not be set back to the parent step. Thus, when the parent step closes, it will
accidentally double-close one of the substeps and raise an exception. This
change causes the "with stream.step()" context manager to aggressively print
STEP_CURSOR annotations so we will make sure we're annotating the correct step
when we close it.

R=iannucci@chromium.org, xusydoc@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15968003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@202149 0039d316-1c4b-4281-b951-d872f2087c98