Fix lookup path handling.
subprocess.Popen takes an env={} argument, which it uses to modify the environment of the child process. However, the PATH variable inside |env| is not used to look up the child process executable -- for that, subprocess.Popen uses os.environ['PATH']. So we need to modify os.environ prior to starting the child process, so it can take advantage of PATH manipulations.

R=iannucci@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18864003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@210655 0039d316-1c4b-4281-b951-d872f2087c98