[chromium_android recipes] Add intrumentation tests recipes.
The internal_dir representation in the chomium_android module
is now a config param of type Path which depends on the
dynamic path '[CHECKOUT]'. Hence, this CL removes the special
case to handle the earlier representation of internal_dir
(which was evaluated as a static value based on path.checkout) in the path module.

BUG=286509

Review URL: https://chromiumcodereview.appspot.com/24530005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@226056 0039d316-1c4b-4281-b951-d872f2087c98