Fix to be able build python's coverage package on Windows.
If we want to be able to build the coverage package on Windows we need to have the VS90COMNTOOLS environment variable set.

This shouldn't necessary point to an installation of VS2008 (at least for this package), this is only used by http://svn.python.org/projects/python/branches/release27-maint/Lib/distutils/msvc9compiler.py to find vcvarsall.bat. We can fake it to point to the version of MSVC checked in to depot_tools.

This mean that we'll use VS2013 compiler and linker, to build this package (while Python27's build with VS2008).

R=iannucci@chromium.org, scottmg@chromium.org

BUG=361796

Review URL: https://codereview.chromium.org/239163004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@264327 0039d316-1c4b-4281-b951-d872f2087c98