Adds unittest support to expect_tests.
  * Supports unittest.TestCase classes and unittest-style modules
    * If methods return data, that becomes the expectation data for that test.
    * We can use this to hybridize some existing tests (e.g. in depot_tools),
      which currently rely on manually curating lists of expectations.
  * Integrates with logging to capture logging lines on test failures

expect_tests tries to break up TestCase's as much as possible, but if they
implement setUpClass or tearDownClass, then they'll be automatically restricted
to run serially on a single subprocess.  Additionally, a TestCase can specify
that it is 'atomic', and so it will be run serially and will not be split up
by the test filter (to support REALLY badly written tests).

R=agable@chromium.org, vadimsh@chromium.org
BUG=366924

Review URL: https://codereview.chromium.org/340963002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@279500 0039d316-1c4b-4281-b951-d872f2087c98