Fix parsing of JSON results when an expected failure flakes.
We had an issue where a test expected to crash started to unexpectedly also timeout occasionally. When we would get the 'TIMEOUT CRASH' result, we would treat the CRASH as an unexpected failure (incorrectly).

Now we explicitly check that the last result was one of the expected results (or one of the 'pass' synonyms like SLOW).

R=ojan@chromium.org, eseidel@chromium.org
BUG=399221

Review URL: https://codereview.chromium.org/446233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@288051 0039d316-1c4b-4281-b951-d872f2087c98