First try of passing properties into functions as arguments.
This CL causes the recipe engine to pass properties to the RunSteps function (or __init__ function for modules) as arguments, if they are both declared in the PROPERTIES top level array, and listed as arguments in the respective function.

These properties support typing, help text, and a default value.

BUG=506174

Review URL: https://codereview.chromium.org/1245023002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296328 0039d316-1c4b-4281-b951-d872f2087c98