Recipe package system.
This provides the infrastructure to create "recipe packages", which are
collections of recipes and modules which can depend on each other.  It
provides a tool recipes.py which interacts with a package.

Corresponding tools/build CL which has a lot of this code reviewed already:
https://codereview.chromium.org/1241323004

BUG=459840

Review URL: https://codereview.chromium.org/1344583003 .