Fix documentation bug for properties in recipes.
This CL fixes the code in the documentation, so that it is valid recipes code.
This is only a temporary fix. The permanent fix involves moving the properties
module inside of the recipe engine, which will end up requiring the buildbot
module, which will contain a bunch of buildbot specific recipe code, which knows
about mastername, buildername, and other buildbot specific concepts.

BUG=532275
R=luqui@chromium.org

Review URL: https://codereview.chromium.org/1346763003 .