Add BoundProperty, an immutable Property, to the recipe engine API.
This CL adds a BoundPropety object to the recipe engine API. The purpose of
this is to avoid mutable data. Because of how we declare properties, the name
of the property is not known by the propety at declaration time. So, we need
to inject this information into the Property.

However, we don't want to mutate the object. We want to strive for properties
being declarative, and allowing mutation on them goes against that. So enter
the BoundProperty. We create a BoundProperty from the regular, unbound
properties, and freeze its data, so that its the final declaration of what
the property should look like.

This was developed because of the discussion in this CL: https://codereview.chromium.org/1343313003

BUG=
R=iannucci@chromium.org, luqui@chromium.org

Review URL: https://codereview.chromium.org/1363773002 .