Add param_name argument to Properties.
This CL adds the param_name member to the Property constructor. This allows
users to have semi name-spaced arguments. This is a temporary fix for now; in
the future, there will be more actual namespacing. Something like
'project.module.property'. For now, this allows you to do something like

  Property('buildbot.foo', param_name="buildbot_foo")

And put it in your module like

  def __init(..., buildbot_foo):
    self._buildbot_foo = buildbot_foo

BUG=
R=luqui@chromium.org

Review URL: https://codereview.chromium.org/1343313003 .