path: implement persistent cache for kitchen path_config on linux
We need to plumb platform name to config for this (the path is going
to be different at least on windows: /b vs. e:\b or even c:\b).

To do that, we need to set the config lazily, so the platform recipe
module can be accessed. It's not available yet in the path module's ctor,
due to how recipe loader works.

BUG=chromium:596887

Review URL: https://codereview.chromium.org/1827753002