Add per step timeout
This CL adds configurable per step timeouts to the step api.
Any call to the step api with a keyword argument of timeout which
takes longer than the given time to run will cause the recipe engine to
throw an exception, which users can optionally handle in their code.

Fixes #3

Review-Url: https://codereview.chromium.org/2062403003