Rework vendoring, imports, packages, protobuf.
This CL primarily targets the protobuf library, which was upgraded from
proto2 to proto3. In doing so, several existing and new issues were
discovered with the way the recipe engine code interacts with the
protobuf library.

"third_party" is no longer a package, to avoid dual-importing its
contents (e.g., "recipe_engine.third_party.google.protobuf" !=
"google.protobuf"). All relative imports to "third_party" packages are
now imports from the "sys.path" space.

In order to handle the "sys.path" and associated import environment,
"recipe_engine/env.py" was introduced. This is a bit more complicated
than the previous import scheme because it contains the necessary hacks to
make it so the new protobuf 3 library load. The previous hacks (including
vendoring "pkg_resource") probably didn't actually work before with proto2,
but because the system library, if one was installed, was also proto2,
nobody noticed or cared.

To manage all this, a "regenerate.py" script was added that regenerates,
from scratch, several vendored items and their derivative files. This will
automatically check out and install various vendored dependencies based on
a pinned configuration, and is self-contained.

BUG=chromium:628770
TEST=None

Review-Url: https://codereview.chromium.org/2137793002