Formally define step config, pass to stream.
Formally define the step configuration dictionary members using a
namedtuple. Replace instances of step dictionary and step dictionary
manipulation with the new "recipe_api.StepConfig".

Pass this new StepConfig struct to the "new_step_stream" stream API call
instead of just some select step properties. The StreamEngine
implementation is responsible for deciding how to process and/or use it.

Similarly define a rendered StepConfig as a RenderedStep. Formally pass
that around to components requiring a rendered value.

This allows for greater control over the specification, manipulation,
and consumption of this internal struct.

Change expectation rendering to refrain from rendering empty StepConfig
parameters. This means that if "cmd" is empty, it will just not appear,
which is more consistent with the intent of the step than an empty list.

BUG=chromium:628770
TEST=local

Review-Url: https://codereview.chromium.org/2253943003