Prefer overridden dependencies.
Currently, when overriding packages, the overridden packages must still
declare the same dependency versions as regular packages. This requires
the user to sync everything prior to experimenting with cross-repository
patches.

Change this behavior to make it so that overridden packages' dependency
set is preferred to the root package's dependencies. This creates a
functional illusion that the overridden packages have been rolled, and
that the root package is being considered candidate for rolling, which
is useful for developers.

BUG=None
TEST=local
  - Ran unit tests.
  - Ran against overridden dependencies with inconsistent dependency
    trees.

Review-Url: https://codereview.chromium.org/2343733003