More strategic retries in fetch.
Define a generic conditional exponential backoff retry decorator.
Reimplement the content in fetch to use this decorator instead of
"_retry". Move decorator usage to the specific site of retryable failure
and only retry known-retryable errors so we don't waste time on failures.

This also adds a retry to Gitiles JSON fetching and a unit test to
assert that this retry logic works successfully.

Defines GitilesFetchError, which includes the status code and error text
ain the output message.

BUG=chromium:649529,chromium:649466
TEST=unit

Review-Url: https://codereview.chromium.org/2362993002