RecipeScript no longer copies recipe global attrs.
The RecipeScript object previously imported the recipe's globals into
its own attribute space. This poses a problem, namely that if the recipe
has any globals that collide with the RecipeScript's members (e.g.,
"name", "run", etc.), the RecipeScript will overwrite them and break.

Instead, RecipeScript is now a pure container for a recipe. It exposes
the recipe's properties and parameters as needed, but doesn't attempt to
act like it is, _itself_, the recipe.

BUG=None
TEST=unit

Review-Url: https://codereview.chromium.org/2417463004