Add clients, requirement injection.
Add clients to recipe engine. A client is a recipe- and recipe-module
facing facet of the recipe runtime.

Replace recipe engine fields that are exported to RecipeApi instances
with recipe engine client instances. This greatly restricts the surface
of recipe module / recipe engine interactions, allowing it to be audited
and documented in a more user-friendly location.

BUG=chromium:628770
TEST=unit

Review-Url: https://codereview.chromium.org/2404233003