Add exit status proto
This CL adds an exit status proto. This exposes more information about how
a recipe execution went, and in particular, if it failed, what kind of a failure
it was. This is needed for our infrastructure to properly distinguish between errors
in the git fetches involved in the recipe engine remote call, and exceptions in
the actual recipe execution.

BUG=646142

Review-Url: https://codereview.chromium.org/2360083008