step: make all args to __call__ explicit
This is a step towards removing overlap between kwargs
and the context object. The goal is for any of the args
to only ever pass it in one way (args or context),
but not in two different, confusing ways.

See https://groups.google.com/a/chromium.org/d/msg/infra-dev/p8Iq9v9Y4k0/w__b1zTWAQAJ
for more context.

BUG=685746

Review-Url: https://codereview.chromium.org/2659693002