[path] Improve abs_to_path to include resource paths.
This introduces a new client to encapsulate the weirdness of the path derivation
outside of the path module and the engine. This client walks the api object
passed to the recipe for every resource path which the recipe might use. It then
allows lookup into this list of paths with a binary search to find the one which
fits the given string-path the best.

It's only intended for the 'path' module (or similar userspace
reimplementation) to use this client. No end-user recipes should ever need it.

R=martiniss@chromium.org, phajdan.jr@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2669303003