raw_io module: Add text_input and text_output
This CLs adds text_input and text_output to the raw io module. These are the
same as input and output, except that they assume the data they handle is
utf-8 encoded, and will decode or encode it as appropriate. If any errors are
encountered while doing this, they will be replaced with an appropriate
replacement character, as defined in https://docs.python.org/2/library/codecs.html#codec-base-classes

BUG=686280,664352

Review-Url: https://codereview.chromium.org/2672593002