[bundle] add mechanism to express precise file dependencies.
This restricts the bundle command, by default, to just the files:
  * recipes/**
  * recipe_modules/**
  * infra/config/recipes.cfg

(minus the expectations). It also allows recipe_modules/** to contain files
named bundle_extra_paths.txt, which contain a list of files, dirs and simple
globs that are required to make the given repo and/or recipe_module work.
The full set of 'extra' paths will be copied into the bundle, allowing even
wacky dependencies (such as the ones that currently exist in the build
repo), to work without modification.

The extra file is a text file to allow the bundling logic to be
language-agnostic (as opposed to putting the file list inside one of the python
files)

R=dnj@chromium.org, phajdan.jr@chromium.org, vadimsh@chromium.org, martiniss@chromium.org
BUG=662654,620765

Review-Url: https://codereview.chromium.org/2702983002