[step_runner] add logic to resolve absolute path of argv[0] on windows.
CreateProcess will automagically append '.exe' if the application
supplied doesn't have one, which explains why "cipd" works on bots
(cipd.exe is in PATH), but not for devs (cipd.bat is in PATH, but no
cipd.exe).

This CL will hunt through PATHEXT for each folder on PATH, which is what
cmd.exe does when you type e.g. 'cipd'. If the supplied command already has
an extension or is an absolute path, behavior will not change.

This hunt is done before printing the step to stdout, so we should be
able to see when this logic had an effect.

R=vadimsh@chromium.org,dnj@chromium.org
BUG=697327

Review-Url: https://codereview.chromium.org/2727553005