[package.proto] add canonical_base_url field.
Also document the fields in package.proto and re-run protoc.

canonical_base_url will be used when generating new Markdown documentation
to allow for meaningful hyperlinks. It's also generally a good idea so when
packages get mirrored, they'll always point to the source of truth.

NOTE: This changes the tag values for the proto definition... however, since
we've never used the binary form of the package message, it shouldn't be a
problem. This allows canonical base url to sort above 'deps'. Otherwise it gets
sorted to the bottom of the file, which is weird.

R=dnj@chromium.org, martiniss@chromium.org, phajdan.jr@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2728303004