[doc] Rewrite doc command to produce structured proto directly from python ast.
This will allow us to:
  * not require executing recipe code when generating documentation.
  * emit real (useful!) markdown documentation (followup CL).
  * emit JSON type/documentation information (potentially for integration with
    cs.chromium.org).

This uses ast.parse in place of loading all the modules. It gathers
documentation from recipe repos, recipe_modules and also recipes, finding module,
function and class docstrings. It understands common core imports of recipe_engine
modules, and will be able to generate hyperlinks to either the recipe_engine
docs or the source code. It collects line numbers and relative paths and so will
be able to hyperlink functions directly to the source in the git repo and/or to
cs.chromium.org links.

This CL implements the new doc proto and data gathering. Markdown generation
will happen in a subsequent CL.

BUG=

Review-Url: https://codereview.chromium.org/2735543003