Use "vpython" to bootstrap the recipe engine.
Remove the current recipe engine bootstrap. Replace it with a
"vpython"-based bootstrap.

The current bootstrap has any number of hacks and problems that have
traditionally caused flake and non-conformity in target systems. The
"vpython" method is believed to have fewer of these problems.

Additionally, replace some vendoring with actual Python wheels. The
vendored packages will remain in place so long as bootstrapping is
optional so that they are accessible from non-bootstrap installations.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2785633003