[url] Cleanup, bugfix, step_test_data.
Add "step_test_data" to the "url" module, allowing default text and JSON
alues to be specified for steps. This is useful to downstream users who
may not want to explicitly specify test data for each step.

Fix a bug in "pycurl.py" where, when stripping prefixes, it applied the
wrong length.

Use a namedtuple to represent the "url" module's status result JSON.
This formally defines the expectation, rather than leaving it up to an
open-ended dict.

BUG=None
TEST=local
  - ./recipes.py test train
  - ./recipes.py run url:example

Review-Url: https://codereview.chromium.org/2881993002