[path] Add concept of a "cleanup" directory.
Some recipe code, such as "bot_update", expect a directory to exist such
that files placed in that directory will be purged in between build
executions. This delayed deletion mechanism is important, as it offers a
facililty for files to be purged, which is potentially time-consuming,
without counting against actual scheduled build time.

BUG=chromium:725631
TEST=expectations
R=hinoka@chromium.org, iannucci@chromium.org

Review-Url: https://codereview.chromium.org/2931893003