[path] Add "base_path" to __contains__, add get.
Currently, the "__contains__" check only returns whether a key is in
"dynamic_paths". If the key is in "base_paths", it returns False.
Unfortunately, this means that different keys are treated differently,
even though users standard users aren't exposed to this difference.

Similary, add a "get" method to retrieve the contents of a path. This
is similar to "dict" "get", and allows for a default value. At the
moment, the only equivalent is to do a try/catch on KeyError, which is
not optimal.

BUG=None
TEST=expectations
R=iannucci@chromium.org

Review-Url: https://codereview.chromium.org/2925373002