Add source manifest proto to recipe_engine.
This adds the initial form of the "source manifest" proto which will
allow recipe_engine to communicate structured data about
stuff-on-disk-that-we-got-from-remote-servers to LUCI services like Milo,
and will also give us a better option than 'got_revision' properties for
exporting this data from recipes (i.e. we'll be able to export all the
data and then query/read it instead of hard-coding a few hacks to export
specific repos).

This only covers git, cipd and isolated for now, but could be expanded
to other data sources, if needed. We don't currently anticipate a need
for that, but the proto is hopefully obviously extensible for this.

Clients consuming this proto (like Milo) may have specialized support
when comparing two manifest (e.g. two entries sharing the same git repo
may render as a git log), but it should also be somewhat easy to show a
useful generic diff when comparing two Manifest protos, even when the
deployment type changes (i.e. local path "foo" used to be a git checkout,
but is now multiple CIPD packages).

R=dnj@chromium.org, hinoka@chromium.org, machenbach@chromium.org
BUG=468053

Review-Url: https://codereview.chromium.org/2998523002