Make cmd argument type assertion error more clear.
Previous error messages only say that the type is incorrect. It doesn't
say what the actual type is. This is usually hard for debugging.

In this cl, I am making it clear what the actual type of the argument
is, in the hope that engineers can spot what is wrong much more quickly.

Bug:726431
Change-Id: If4b165112c94c9a3bb380003127c52530c00e27d
Reviewed-on: https://chromium-review.googlesource.com/835529
Commit-Queue: Zhiling Huang <hzl@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>