[raw_io] Make it succeed even with read only files.
Use a simplified algorithm compared to swarming's client file_path.py, as we
don't expect the steps here to be as egregious in the file ACL than what
Swarming tasks can do.

Tested manually on Windows by copy pasting the code, creating a directory and a
file in it, running 'attrib +r' on the file. Testing with and without the
onerror handler.

Bug: 846383
Change-Id: I13181eb381edb360cbbe69d95c36a266ffb8bb03
Reviewed-on: https://chromium-review.googlesource.com/1072068
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>