Add simple `archive` module for tar + zip.
This adds a very simple pure-python archive file module to the base
recipe_engine. This will replace all similar simple archive modules in
downstream repos.

This is needed to add an `archive` endpoint function to the gitiles
module in depot_tools, which, in turn, is needed to deprecate the
recipes-py builder on buildbot.

R=phosek@chromium.org, tandrii@chromium.org, vadimsh@chromium.org

Bug: 790650
Change-Id: Ie68d3f93ab7c182d86b4303f0ff4680e60ac6d15
Reviewed-on: https://chromium-review.googlesource.com/1115736
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>