Enable adding an output log when using raw_io.output_text.
Currently v8/api.py updates the presentation of a step created by
chromium.run_mb to add a log containing the content of the
raw_io.output_text that was provided as the stdout for the step. This
prevents modifying the implementation of chromium.run_mb to include
additional steps that provide quality-of-life improvements. Adding the
add_output_log parameter enables accomplishing the same thing without
relying on the implementation detail of chromium.run_mb having only a
single step.

Change-Id: I5a6573b4d95904735f4ccd7fd5932de484d4c56b
Bug: 649209
Reviewed-on: https://chromium-review.googlesource.com/1158802
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>