Keep fewer copies of data to sanitize utf-8 in OutputPlaceholders too.
This does for OutputPlaceholders what https://chromium-review.googlesource.com/c/1181246/
did for InputPlaceholders. In the other CL, we had a string that was maybe utf-8
and wanted to write it to a file.

Here we have a file that maybe contains utf-8 and want to convert it to a string.

We can read the file in chunks so we don't have to keep the whole contents
in memory, but we then need to combine all the converted strings to one full
string. That requires one copy, so this only brings us from 400% memory overhead
to around 100% overhead. Still much better, but not as much better as the
other change.

With a larger change in recipes (e.g. by making result() return a callable
instead of a string) larger savings are possible, but let's go with the
smaller patch for now.

Bug: chromium:735306
Change-Id: I8eaac86d2526119c28a61dd36cf244b32a30950c
Reviewed-on: https://chromium-review.googlesource.com/1184059
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>