[file] Add flatten_single_directories function.
This is useful when you have a bunch of files/dirs at the end of a long
string of directories. Instead of hard-coding the intermediate directories,
you can use this function to move them all up to the top folder (and then
delete the intermediate dirs).

An immediate application of this is for tarballs e.g. `foo-1.2.tar.gz`
containing all of its files in a subdir `foo-1.2`.

This CL also deletes a bogus `example.disable` file, as well as some
minor pylint nits.

R=tandrii@chromium.org

Change-Id: I2a1fb79d52b940b23bd9a5698db1cf95c92fe793
Reviewed-on: https://chromium-review.googlesource.com/1186138
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>