Add a sentinel object for representing all return codes are ok.
Currently, when if ok_ret is set to 'any' or 'all' when using the step
API, it's transformed into range(-256, 256) which results in every
value in that range being dumped into the stdout of step. The use of a
sentinel instead results in output of the form "ok_ret: ALL_OK".

Bug: 671950
Change-Id: Iee661b4bb8f34cc27da522b0626c4656e21aac1a
Reviewed-on: https://chromium-review.googlesource.com/1195592
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>