[cipd] Add better describe simulation behavior.
This will make the `describe` method have a non-zero exit code and a
proper-looking error message in the event that the describe command is
told to return NO results (no tags and no refs). This allows the correct
simulation of `cipd describe` commands for versions which don't exist.

R=dnj@chromium.org, hinoka@chromium.org, tandrii@chromium.org

Bug: 877775
Change-Id: Ida8757700e100fa267abbef62ee2f252ab225446
Reviewed-on: https://chromium-review.googlesource.com/1205641
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>