[buildbucket] Pass v2 build as a dict, not base64 bytes
"$recipes-py/buildbucket".build is base64-encoded bytes of build message.
It does the job of discouraging recipe authors from interpeting it,
but at the same time makes it very difficult to specify it for tools
that specify input properties, e.g. local run of recipes.py or led.

Make it a dict that represents a JSONPB-encoded protobuf message.
When parsing, ignore unrecognized fields so that adding a new field
in the proto does not break prod jobs.

Change-Id: I14e9bec61431eb09fe9ce38c1f1a9e7369197f13
Reviewed-on: https://chromium-review.googlesource.com/1222232
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>