Improve local execution of recipes
This CL adds the --annotator-output flag, which determines if recipes
should output annotations when executing. By default, this is false
if sys.stdout is terminal. This means that, when executing recipes
locally, the output should be much more readable.

It also adds a --log-temp-dir flag. If set, the recipe engine will dump
any large output streams into a file in that directory. This also helps
the output to be more readable.

Change-Id: Ib215212a61caee9ebf458ed788a8e831b17557f8
Reviewed-on: https://chromium-review.googlesource.com/1213685
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>