[StepPresentation] Remove all logs from memory during finalization.
Previously the recipe engine kept all log content in memory for
the duration of the recipe execution. No recipes have been found
that actually need to read this data (which can't read it in some
alternate mechanism). Since the logs can be quite long, this should
reduce the memory consumption of most recipe executions, sometimes
by a lot.

Because the logs are no longer kept in memory, the recipe engine
will now pop an assertion if a recipe attempts to read/write a log
after the step has been finalized.

R=martiniss@chromium.org, nodir@chromium.org

Bug: 833539,735306
Change-Id: I8450b473eeaaafcdc0a87c7c542f961848327050
Reviewed-on: https://chromium-review.googlesource.com/c/1171206
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>