[doc/user_guide.md] Initial structure for rewrite of user_guide.md.
The existing guide is... not helpful. I'm rewriting the guide to be
better organized and more useful.

Obviously there's a lot that needs to be filled in, but I wanted to
submit this first so that I have a structure to work with, and can
send smallish incremental CLs.

R=martiniss@chromium.org, tandrii@chromium.org

Change-Id: I696c9e17e0b41fe9b579912a2f6c781696e8111f
Reviewed-on: https://chromium-review.googlesource.com/c/1325390
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>