[buildbucket] Add set_output_gitiles_commit
Add a recipe function to set output gitiles commit of a build.
It is currently implemented as an output property with a dummy step, because
there is currently no better way to propagate this to kitchen.
Most likely the impl will be replaced when recipe engine speaks build proto,
but the public API will remain the same.

R=iannucci@chromium.org

Bug: 851585
Change-Id: I7477919c5e8e1d7ca2f319774af89f2ebbb9a89d
Reviewed-on: https://chromium-review.googlesource.com/c/1390771
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>