[isolated] Allow ability to isolate a given root directory.
This fixes a bug in which an Isolated object with a fixed root
directory could not reflexively archive that directory.

Also, this change makes it so an Isolated object is initialized with
IsolatedApi.m instead of an instance of IsolatedApi itself, as then the
private member _api is more deserving of its name, in analogy with the
`api` that is passed to RunTests().

Bug: 919330
Change-Id: I79711810e4dcbe8cfb449af27afec7e9ac5e9970
Reviewed-on: https://chromium-review.googlesource.com/c/1396695
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>