[depgraph] Remove depgraph subcommand for now.
I'm currently rewriting the loader/package logic to be simpler and
more correct; the depgraph subcommand currently relies heavily on the
current interfaces there (and so needs to be updated).

This command is theoretically useful, however it currently produces very
difficult-to-use output. Updating the depgraph command for the new loader
code constitutes, essentially, a rewrite, and if I'm going to have to
rewrite this command, I'd rather spend the time writing a version which
had useful/actionable output.

The original design of the command was fine for when it was introduced,
but today the complexity of the recipe dependency graph has increased so
much to the point where this graph appoach no longer works.

R=jchinlee@chromium.org, martiniss@chromium.org, tandrii@chromium.org

Bug: 914992, 909848
Change-Id: I6e69fb64145ef0e11bcebd9e1957687b7bb3026a
Reviewed-on: https://chromium-review.googlesource.com/c/1441632
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>