Add support for longMessage and maxDiff to assertions api.
The longMessage attribute changes the behavior when an assertion fails
and a custom message is provided so that the standard failure message is
displayed followed by the custom message, it can be helpful when making
multiple assertions to use the custom message as an identifier for the
assertion so that similar failures can be distinguished from one
another.

The maxDiff attribute controls the behavior when an assertion fails on
some comparison types (e.g. assertEquals when comparing lists). When an
assertion failure occurs, a diff of the data is displayed so that the
whole state of the values being tested can be seen. The diff is limited
in size by the maxDiff attribute, with a value of None indicating no
limit.

Change-Id: I7a108b5415e5e89c048708a6e5887f871d710e55
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1509972
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>